% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcall.R
\name{eval_ggcall}
\alias{eval_ggcall}
\title{Evaluate ggcall}
\usage{
eval_ggcall(call, ...)
}
\arguments{
\item{call}{An expression representing the ggplot construction call, typically
generated by `ggcall()`. This expression should have an
attribute 'ggcall_env' that stores the environment in which
the plot was originally created.}

\item{...}{Additional variables passed to the evaluation environment.}
}
\value{
The resulting ggplot object produced by evaluating the expression `x`.
}
\description{
This function evaluates an expression representing a ggplot construction code.
It specifically uses the environment stored in the 'ggcall_env' attribute
of the expression, ensuring that the plot is reconstructed in the correct context.
}
\examples{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()
plot_call <- ggcall(p)
reconstructed_plot <- eval_ggcall(plot_call)
print(reconstructed_plot)

}
