% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{eval_ggplot_code}
\alias{eval_ggplot_code}
\title{Evaluate Plot Construction Code}
\usage{
eval_ggplot_code(code)
}
\arguments{
\item{code}{An expression representing the ggplot construction code, typically
generated by `get_ggplot_code`. This expression should have an
attribute 'plot_history_env' that stores the environment in which
the plot was originally created.}
}
\value{
The resulting ggplot object produced by evaluating the expression `x`.
}
\description{
This function evaluates an expression representing a ggplot construction code.
It specifically uses the environment stored in the 'plot_history_env' attribute
of the expression, ensuring that the plot is reconstructed in the correct context.
}
\examples{
p <- ggplot(mtcars, aes(x=wt, y=mpg)) + geom_point()
plot_code <- get_ggplot_code(p)
reconstructed_plot <- eval_ggplot_code(plot_code)
print(reconstructed_plot)

}
