% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{get_ggplot_code}
\alias{get_ggplot_code}
\title{Retrieve Plot Construction Code from ggplot_history Object}
\usage{
get_ggplot_code(plot, call = TRUE)
}
\arguments{
\item{plot}{A ggplot object of class 'ggplot_history'.}

\item{call}{Logical; if TRUE, returns a callable expression representing the
plot construction. If FALSE, returns a list of the history elements.}
}
\value{
Depending on the value of 'call', either a callable expression or
a list representing the history of the ggplot object.
}
\description{
Extracts the complete history of a ggplot object's construction,
providing a way to reproduce or inspect the plot. Designed to work
with ggplot objects of class 'ggplot_history'.
}
\examples{
p <- ggplot(mtcars, aes(x=wt, y=mpg)) + geom_point()
plot_code <- get_ggplot_code(p)
print(plot_code)

# eval(plot_code) # Reproduce the plot

}
