% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-patchwork.R
\name{-.ggcall}
\alias{-.ggcall}
\alias{/.ggcall}
\alias{|.ggcall}
\alias{*.ggcall}
\alias{&.ggcall}
\title{Minus Operator for ggcall Objects}
\usage{
\method{-}{ggcall}(e1, e2)

\method{/}{ggcall}(e1, e2)

\method{|}{ggcall}(e1, e2)

\method{*}{ggcall}(e1, e2)

\method{&}{ggcall}(e1, e2)
}
\arguments{
\item{e1}{The left-hand side `ggcall` object.}

\item{e2}{The right-hand side `ggcall` object.}
}
\value{
A combined `ggcall` object after applying the minus operation.

A combined `ggcall` object after applying the division operation.

A combined `ggcall` object after applying the or operation.

A combined `ggcall` object after applying the multiplication operation.

A combined `ggcall` object after applying the and operation.
}
\description{
Applies the minus operator to `ggcall` objects, utilizing the corresponding operator from the `patchwork` package.

Applies the division operator to `ggcall` objects, utilizing the corresponding operator from the `patchwork` package.

Applies the or operator to `ggcall` objects, utilizing the corresponding operator from the `patchwork` package.

Applies the multiplication operator to `ggcall` objects, utilizing the corresponding operator from the `patchwork` package.

Applies the and operator to `ggcall` objects, utilizing the corresponding operator from the `patchwork` package.
}
\details{
This function allows for the subtraction of `ggcall` objects using the `patchwork` syntax.

This function allows for the division of `ggcall` objects using the `patchwork` syntax.

This function allows for the combination of `ggcall` objects using the `patchwork` syntax.

This function allows for the multiplication of `ggcall` objects using the `patchwork` syntax.

This function allows for the combination of `ggcall` objects using the `patchwork` syntax.
}
\keyword{internal}
