% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-ggcall.R
\name{ggcall_env}
\alias{ggcall_env}
\title{Retrieve Environment from ggcall}
\usage{
ggcall_env(call)
}
\arguments{
\item{call}{An expression representing the ggplot construction code.}
}
\value{
The environment in which the ggplot construction code was created.
}
\description{
Extracts the environment in which the ggplot construction code
was originally created. This function is designed to work with expressions
generated by `ggcall`.
}
\examples{
fun <- function(data, x, y) {
  ggplot(data, aes(x = !!as.name(x), y = !!as.name(y))) +
    geom_point()
}
gplot <- fun(mtcars, "wt", "mpg")
plot_call <- ggcall(gplot)
env <- ggcall_env(plot_call)
ls(env)
env[["data"]]
as.list(env)
}
