% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-ggcall.R
\name{ggplot}
\alias{ggplot}
\title{Enhanced `ggplot` Function with History Tracking}
\usage{
ggplot(...)
}
\arguments{
\item{...}{Arguments passed to the original ggplot function from ggplot2.}
}
\value{
A ggplot object of class 'ggcall', with an additional
attribute 'ggcall' that stores the history of plot construction.
}
\description{
Overrides the default `ggplot` function from the ggplot2 package, adding the
capability to track the history of plot construction.
This function initializes a history attribute in the `ggplot` object.
}
\examples{
library(ggplot2, exclude = "ggplot")
p <- ggplot(mtcars, aes(x = wt, y = mpg))
# the + function has to come from ggcall package
attr(p + geom_point(), "ggcall")

}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
