% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.cluster.R
\name{marker.cluster}
\alias{marker.cluster}
\title{marker.cluster}
\usage{
marker.cluster(marker, data.type = "raw", iteration = 5,
  model.list = c("Gaussian_pk_Lk_Ck", "Gaussian_pk_Lk_Bk"), clust.num = 8)
}
\arguments{
\item{marker}{A project object, must be formatted correctly according to the project.create function}

\item{data.type}{A string variable indicating which data should be used for clustering. ('raw', 'GSnorm','GStrans', 'thetaonly')}

\item{iteration}{The number of times to re-cluster (can result in more stable results, but slows clustering time)}

\item{model.list}{A list of Rmixmod models to use for cluster (Refer to Rmixmod documentation for more info)}

\item{clust.num}{The maximum number of clusters to test for (lowering this speeds clustering time)}
}
\value{
A marker object updated with the best clustering result out of a defined number of iterations

This function relies on the following packages:
Rmixmod
}
\description{
Clusters a single markers with Rmixmod
}
\details{
USes the 'Rmixmod' package to fit a Gaussian mixture model using a supplied model list
}
