% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.create.R
\name{marker.create}
\alias{marker.create}
\title{marker.create}
\usage{
marker.create(name = "NA", x = vector(), y = vector(),
  x.scaled = vector(), y.scaled = vector(), R = vector(),
  theta = vector(), theta.trans = vector(), mixmodout = NULL,
  miss.ind = NULL, clust.time = NULL, data.type = NULL)
}
\arguments{
\item{name}{the marker ID}

\item{x}{a vector of x.values}

\item{y}{a vector of y.values}

\item{x.scaled}{a vector of scaled x.values}

\item{y.scaled}{a vector of scaled y.values}

\item{R}{a vector of R values (either from GenomeStudio or calculated from raw values)}

\item{theta}{a vector of theta values (either from GenomeStudio or calculated from raw values)}

\item{theta.trans}{a vector of arcsin transformed values}

\item{mixmodout}{cluster results from mixmodCluster}

\item{miss.ind}{a vecor of missing value indices}

\item{clust.time}{timing results for the clustering}

\item{data.type}{Which data points were used for clustering}
}
\value{
An object storing all data for a single marker
}
\description{
Creates an empty Marker object
}
