% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.plot.R
\name{marker.plot}
\alias{marker.plot}
\title{marker.plot}
\usage{
marker.plot(marker, recast = F)
}
\arguments{
\item{marker}{A marker object, must have a valid 'mixmodout' cluster model#'}

\item{recast}{Should}
}
\value{
An object storing all data for a single marker

This function relies on the following packages:
}
\description{
Plots the clustering results for a marker object. Includes elipses to indicate distribution variance/rotation.
}
