% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-patchwork.R
\name{patch_operator_base}
\alias{patch_operator_base}
\title{Base Function for Patchwork Operators}
\usage{
patch_operator_base(e1, e2, operator, class)
}
\arguments{
\item{e1}{The left-hand side `ggcall` object.}

\item{e2}{The right-hand side `ggcall` object.}

\item{operator}{The operator as a string (e.g., "-", "/", "|", "*", "&").}

\item{class}{The class to which the operator is applied ("ggplot" or "gg").}
}
\value{
A combined `ggcall` object resulting from the operation.
}
\description{
A helper function that applies patchwork operators to `ggcall` objects.
}
\keyword{internal}
