% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcall.R
\name{+.gg}
\alias{+.gg}
\title{Custom '+' Operator for ggcall Objects}
\usage{
\method{+}{gg}(e1, e2)
}
\arguments{
\item{e1}{A ggplot object of class 'ggcall'.}

\item{e2}{A layer or theme to add to the ggplot object.}
}
\value{
A modified ggplot object with updated plot history.
}
\description{
Enhances the '+' operator for ggplot objects to track the history of
plot layers and modifications. This function is meant to be used in
conjunction with the enhanced ggplot function provided by this package.
}
\examples{
p <- ggplot(mtcars, aes(x=wt, y=mpg)) + geom_point()
attr(p, "ggcall") # View the plot call

}
