% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.cluster.R
\name{project.cluster}
\alias{project.cluster}
\title{project.cluster}
\usage{
project.cluster(project, marker.indices = NULL, data.type = "raw.scaled",
  iteration = 5, model.list = c("Gaussian_pk_L_Ck", "Gaussian_pk_L_Bk"),
  clust.num = 8)
}
\arguments{
\item{project}{A project object, must be formatted correctly according to the project.create function}

\item{marker.indices}{The indices of the markers to cluster}

\item{data.type}{A string variable indicating which data should be used for clustering. ('raw', 'GSnorm','GStrans', 'thetaonly')}

\item{iteration}{The number of times to re-cluster (can result in more stable results, but increases clustering time)}

\item{model.list}{A list of Rmixmod models to use for cluster (Refer to Rmixmod documentation for more info)}

\item{clust.num}{The maximum number of clusters to test for (lowering this speeds clustering time)}
}
\value{
A project object updated with clustering models

This function relies on the following packages:
}
\description{
Cluster markers in Parallel with Rmixmod
}
\details{
Sets up a parallel backend and clusters each marker using mixture of gaussian models provided by the 'Rmixmod' package.
}
