% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.genotype.R
\name{project.genotype}
\alias{project.genotype}
\title{project.genotype}
\usage{
project.genotype(project, format = "none")
}
\arguments{
\item{project}{A project object, must contain cluster information from the project.cluster function}

\item{format}{If clusters were called using raw intensities or scaled raw intensities, should they be transformed to theta an R? (can be 'none' or 'transformed')}
}
\value{
A list of genotype information: 3 matrices corresponding to the cluster, the x/theta cluster center, the y/R cluster center, the specified format, and the data type used for clustering
}
\description{
Uses Rmixmod clusters from a GGcall project to call genotypes
}
