% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.import.R
\name{project.import}
\alias{project.import}
\title{project.import}
\usage{
project.import(exportpath = NULL, raw = T, GSnorm = T)
}
\arguments{
\item{exportpath}{the path to the GenomeStudio .txt data export}

\item{raw}{Logical T/F: Whether to import raw data}

\item{GSnorm}{Logical T/F: Whether to import GenomeStudio transformed/normalized data}
}
\value{
A project object containing a vector of marker names and a list of marker objects
}
\description{
Imports a Genome Studio data export file to only import useful information for clustering (theta, R, raw x, and raw y)
}
