% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.plot.R
\name{project.plot}
\alias{project.plot}
\title{project.plot}
\usage{
project.plot(project, indices = NULL, save.loc = getwd(), recast = F)
}
\arguments{
\item{project}{a project object, markers should already be clustered}

\item{indices}{A numerical index of the markers to plot}

\item{save.loc}{A string path to the folder to save the images}

\item{recast}{Logical, should raw data be recast to polar theta and r coordinates?}
}
\value{
Nothing

This function relies on the following packages:
ggplot2, car, RColorBrewer, stringr
}
\description{
Plots the clustering results for a marker object. Includes elipses to indicate distribution variance/rotation.
}
