% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcall.R
\name{ggcall}
\alias{ggcall}
\title{Retrieve Construction Call from ggplot Object}
\usage{
ggcall(plot)
}
\arguments{
\item{plot}{A ggplot object of class 'ggcall'.}
}
\value{
Depending on the value of 'call', either a callable expression or
a list representing the history of the ggplot object.
}
\description{
Extracts the complete history of a ggplot object's construction,
providing a way to reproduce or inspect the plot. Designed to work
with ggplot objects of class 'ggcall'.
}
\examples{
p <- ggplot(mtcars, aes(x=wt, y=mpg)) + geom_point()
plot_call <- ggcall(p)
# Optionally: Style the code with styler
styler::style_text(backports:::deparse1(plot_call))

}
