% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.create.R
\name{marker.create}
\alias{marker.create}
\title{marker.create}
\usage{
marker.create(name = "NA", x = vector(), y = vector(),
  x.scaled = vector(), y.scaled = vector(), R = vector(),
  theta = vector(), theta.trans = vector(), mixmodout = NULL,
  miss.ind = NULL, clust.time = NULL, data.type = NULL)
}
\arguments{
\item{name}{the marker ID}

\item{x}{a vector of x.values}

\item{y}{a vector of y.values}

\item{mixmodout}{cluster results from mixmodCluster}

\item{miss.ind}{a vecor of missing value indices}

\item{clust.time}{timing results for the clustering}

\item{x.adj}{a vector of adjusted x.values}

\item{y.adj}{a vector of adjusted y.values}
}
\value{
An object storing all data for a single marker
}
\description{
Creates an empty Marker object
}
