% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.call.R
\name{project.call}
\alias{project.call}
\title{project.call}
\usage{
project.call(project, call.type = "cat.all", min.prop = 0.01)
}
\arguments{
\item{project}{A project object, must be formatted correctly according to the project.create function}

\item{call.type}{A string variable indicating how to call clusters ('cat.all' = All markers w/ indicator expansion)}

\item{min.prop}{The minimum proportion of the population for a cluster to be called}
}
\value{
A marker object updated with the best clustering result out of a defined number of iterations

This function relies on the following packages:
Rmixmod
}
\description{
Calls genotypes for all markers in a project
}
\details{
USes the 'Rmixmod' package to fit a Gaussian mixture model using a supplied model list
}
